TASK: Improve label
Why print one when you can print several. What I did in
paint.exe was to print it once, measure it in mm. Divide
that by the 51mm length of the AVR and then set that as
the scaling percentage for printing - then they come out
at 51mm long which fits nicely.